started gems scripting
