+++ characters.json
@@ -348,7 +348,7 @@
     "rating": 3,
     "featured": false,
     "id": 50
-  },
+  }
  
 ]
