+++ test.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+
+<body style="background-color: gray;">
+    <button onclick="addAmount(100)">ADD 100</button>
+    <button onclick="subtractAmount(100)">SUB 100</button>
+
+</body>
+<script src="gems.js"></script>
+
+</html>
